# docs/openapi.yaml
openapi: 3.0.4
info:
  title: TenX Empires API
  description: REST API for TenX Empires - a turn-based strategy game (Version 1.0)
  contact:
    name: TenX Empires
    url: https://github.com/tenxempires
  version: v1
paths:
...
  /v1/auth/csrf:
    get:
      tags:
      - v1
      summary: Issues or refreshes the CSRF token cookie for the SPA.
      description: >-
        Sets a non-HttpOnly XSRF-TOKEN cookie that the client should echo via the X-XSRF-TOKEN header

        on subsequent non-GET requests protected with ValidateAntiForgeryToken.
      responses:
        204:
          description: Token issued via Set-Cookie (no response body).
        429:
          description: Too many requests (rate limited).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
              example:
                code: RATE_LIMIT_EXCEEDED
                message: Too many requests. Please try again later.
        500:
          description: Failed to issue token.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
              example:
                code: CSRF_ISSUE_FAILED
                message: Unable to issue CSRF token.
...
  /v1/auth/me:
    get:
      tags:
      - v1
      summary: Returns information about the current authenticated user.
      responses:
        200:
          description: Returns the current user info.
        401:
          description: Unauthorized - user is not authenticated.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
              example:
                code: UNAUTHORIZED
                message: User ID claim not found. User must be authenticated.
  /v1/auth/register:
    post:
      tags:
      - v1
      summary: Registers a new account and signs the user in.
      requestBody:
        content:
          application/json; version=1.0:
            schema:
              $ref: '#/components/schemas/RegisterRequestDto'
          text/json; version=1.0:
            schema:
              $ref: '#/components/schemas/RegisterRequestDto'
          application/*+json; version=1.0:
            schema:
              $ref: '#/components/schemas/RegisterRequestDto'
      responses:
        204:
          description: Registered and signed in.
        400:
          description: Invalid input or user already exists.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
              example:
                code: INVALID_INPUT
                message: One or more validation errors occurred.
  /v1/auth/login:
    post:
      tags:
      - v1
      summary: Authenticates a user with email and password.
      requestBody:
        content:
          application/json; version=1.0:
            schema:
              $ref: '#/components/schemas/LoginRequestDto'
          text/json; version=1.0:
            schema:
              $ref: '#/components/schemas/LoginRequestDto'
          application/*+json; version=1.0:
            schema:
              $ref: '#/components/schemas/LoginRequestDto'
      responses:
        204:
          description: Signed in.
        400:
          description: Invalid credentials.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
              example:
                code: INVALID_INPUT
                message: One or more validation errors occurred.
  /v1/auth/logout:
    post:
      tags:
      - v1
      summary: Signs the current user out.
      responses:
        204:
          description: Signed out.
        401:
          description: Unauthorized.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
              example:
                code: UNAUTHORIZED
                message: User ID claim not found. User must be authenticated.
  /v1/auth/forgot-password:
    post:
      tags:
      - v1
      summary: Requests a password reset email for the specified account.
      description: >-
        Returns a generic success response regardless of whether the email exists to prevent account enumeration.

        If the email is associated with an account, a password reset link will be sent.
      requestBody:
        content:
          application/json; version=1.0:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequestDto'
          text/json; version=1.0:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequestDto'
          application/*+json; version=1.0:
            schema:
              $ref: '#/components/schemas/ForgotPasswordRequestDto'
      responses:
        204:
          description: Request processed (email sent if account exists).
        400:
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
              example:
                code: INVALID_INPUT
                message: One or more validation errors occurred.
        429:
          description: Too many requests (rate limited).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
              example:
                code: RATE_LIMIT_EXCEEDED
                message: Too many requests. Please try again later.
  /v1/auth/resend-verification:
    post:
      tags:
      - v1
      summary: Resends the email verification link to the user's email address.
      description: >-
        If authenticated, uses the current user's email. If not authenticated and email is provided,

        sends verification email to that address. Returns success regardless of whether account exists

        to prevent account enumeration.
      requestBody:
        content:
          application/json; version=1.0:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequestDto'
          text/json; version=1.0:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequestDto'
          application/*+json; version=1.0:
            schema:
              $ref: '#/components/schemas/ResendVerificationRequestDto'
      responses:
        204:
          description: Verification email sent (if account exists and is unverified).
        400:
          description: Invalid input.
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
              example:
                code: INVALID_INPUT
                message: One or more validation errors occurred.
        429:
          description: Too many requests (rate limited).
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ApiErrorDto'
              example:
                code: RATE_LIMIT_EXCEEDED
                message: Too many requests. Please try again later.
...
components:
  schemas:
...
    ApiErrorDto:
      type: object
      properties:
        code:
          type: string
          nullable: true
        message:
          type: string
          nullable: true
      additionalProperties: false
...
    ForgotPasswordRequestDto:
      type: object
      properties:
        email:
          type: string
          nullable: true
      additionalProperties: false
...
    LoginRequestDto:
      type: object
      properties:
        email:
          type: string
          nullable: true
        password:
          type: string
          nullable: true
        rememberMe:
          type: boolean
      additionalProperties: false
...
    RegisterRequestDto:
      required:
      - confirm
      - email
      - password
      type: object
      properties:
        email:
          minLength: 1
          type: string
          format: email
        password:
          minLength: 8
          type: string
        confirm:
          minLength: 8
          type: string
      additionalProperties: false
    ResendVerificationRequestDto:
      type: object
      properties:
        email:
          type: string
          nullable: true
      additionalProperties: false
...